chore(utils): move utils to their own files
