+++ src/lib/utils/capitalize.ts
@@ -1,0 +1,10 @@
+function capitalize(string: string) {
+  return (
+    string
+      .toString()
+      .charAt(0)
+      .toUpperCase() + string.toString().slice(1)
+  );
+}
+
+export default capitalize;
