+++ src/lib/utils/escapeRefinement.js
@@ -1,0 +1,9 @@
+function escapeRefinement(value) {
+  if (typeof value === 'number' && value < 0) {
+    value = String(value).replace(/^-/, '\\-');
+  }
+
+  return value;
+}
+
+export default escapeRefinement;
