+++ src/lib/utils/getPropertyByPath.ts
@@ -1,0 +1,7 @@
+function getPropertyByPath(object: object, path: string) {
+  const parts = path.split('.');
+
+  return parts.reduce((current, key) => current && current[key], object);
+}
+
+export default getPropertyByPath;
