+++ src/lib/utils/index.ts
@@ -1,0 +1,46 @@
+import capitalize from './capitalize';
+import isDomElement from './isDomElement';
+import getContainerNode from './getContainerNode';
+import isSpecialClick from './isSpecialClick';
+import bemHelper from './bemHelper';
+import prepareTemplateProps from './prepareTemplateProps';
+import renderTemplate from './renderTemplate';
+import getRefinements from './getRefinements';
+import clearRefinements from './clearRefinements';
+import prefixKeys from './prefixKeys';
+import escapeRefinement from './escapeRefinement';
+import unescapeRefinement from './unescapeRefinement';
+import checkRendering from './checkRendering';
+import getPropertyByPath from './getPropertyByPath';
+import { warning, deprecate } from './log';
+import {
+  createDocumentationLink,
+  createDocumentationMessageGenerator,
+} from './documentation';
+import {
+  aroundLatLngToPosition,
+  insideBoundingBoxToBoundingBox,
+} from './geo-search';
+
+export {
+  capitalize,
+  getContainerNode,
+  bemHelper,
+  prepareTemplateProps,
+  renderTemplate,
+  isSpecialClick,
+  isDomElement,
+  getRefinements,
+  clearRefinements,
+  aroundLatLngToPosition,
+  insideBoundingBoxToBoundingBox,
+  getPropertyByPath,
+  prefixKeys,
+  escapeRefinement,
+  unescapeRefinement,
+  checkRendering,
+  createDocumentationLink,
+  createDocumentationMessageGenerator,
+  deprecate,
+  warning,
+};
