+++ src/lib/utils/isDomElement.js
@@ -1,0 +1,8 @@
+function isDomElement(object) {
+  return (
+    object instanceof window.HTMLElement ||
+    (Boolean(object) && object.nodeType > 0)
+  );
+}
+
+export default isDomElement;
