+++ src/lib/utils/isSpecialClick.ts
@@ -1,0 +1,13 @@
+function isSpecialClick(event: MouseEvent) {
+  const isMiddleClick = event.button === 1;
+
+  return (
+    isMiddleClick ||
+    event.altKey ||
+    event.ctrlKey ||
+    event.metaKey ||
+    event.shiftKey
+  );
+}
+
+export default isSpecialClick;
