+++ src/lib/utils/renderTemplate.js
@@ -1,0 +1,58 @@
+import mapValues from 'lodash/mapValues';
+import curry from 'lodash/curry';
+import hogan from 'hogan.js';
+
+// We add all our template helper methods to the template as lambdas. Note
+// that lambdas in Mustache are supposed to accept a second argument of
+// `render` to get the rendered value, not the literal `{{value}}`. But
+// this is currently broken (see https://github.com/twitter/hogan.js/issues/222).
+function transformHelpersToHogan(helpers, compileOptions, data) {
+  return mapValues(helpers, method =>
+    curry(function(text) {
+      const render = value => hogan.compile(value, compileOptions).render(this);
+      return method.call(data, text, render);
+    })
+  );
+}
+
+function renderTemplate({
+  templates,
+  templateKey,
+  compileOptions,
+  helpers,
+  data,
+}) {
+  const template = templates[templateKey];
+  const templateType = typeof template;
+  const isTemplateString = templateType === 'string';
+  const isTemplateFunction = templateType === 'function';
+
+  if (!isTemplateString && !isTemplateFunction) {
+    throw new Error(
+      `Template must be 'string' or 'function', was '${templateType}' (key: ${templateKey})`
+    );
+  }
+
+  if (isTemplateFunction) {
+    return template(data);
+  }
+
+  const transformedHelpers = transformHelpersToHogan(
+    helpers,
+    compileOptions,
+    data
+  );
+
+  return hogan
+    .compile(template, compileOptions)
+    .render({
+      ...data,
+      helpers: transformedHelpers,
+    })
+    .replace(/[ \n\r\t\f\xA0]+/g, spaces =>
+      spaces.replace(/(^|\xA0+)[^\xA0]+/g, '$1 ')
+    )
+    .trim();
+}
+
+export default renderTemplate;
