+++ src/lib/utils/unescapeRefinement.ts
@@ -1,0 +1,5 @@
+function unescapeRefinement(value: any) {
+  return String(value).replace(/^\\-/, '-');
+}
+
+export default unescapeRefinement;
