+++ src/widgets/pagination/__tests__/pagination-test.js
@@ -1,5 +1,5 @@
 import { render } from 'preact-compat';
-import { getContainerNode } from '../../../lib/utils';
+import getContainerNode from '../../../lib/utils/getContainerNode';
 import pagination from '../pagination';
 
 jest.mock('preact-compat', () => {
@@ -10,11 +10,11 @@ jest.mock('preact-compat', () => {
   return module;
 });
 
-jest.mock('../../../lib/utils', () => {
-  const module = require.requireActual('../../../lib/utils');
+jest.mock('../../../lib/utils/getContainerNode', () => {
+  const module = require.requireActual('../../../lib/utils/getContainerNode');
 
-  const _getContainerNode = module.getContainerNode;
-  module.getContainerNode = jest.fn((...args) => _getContainerNode(...args));
+  const _getContainerNode = module.default;
+  module.default = jest.fn((...args) => _getContainerNode(...args));
 
   return module;
 });
